+++ test/src/test/java/com/devexperts/timetest/test/Repeat.java
@@ -1,4 +1,4 @@
-package com.devexperts.timetest;
+package com.devexperts.timetest.test;
 
 /*
  * #%L