+++ transformer/src/main/java/com/devexperts/timetest/transformer/GlobUtil.java
@@ -6,7 +6,7 @@
  * If a copy of the MPL was not distributed with this file, You can obtain one at
  * http://mozilla.org/MPL/2.0/.
  */
-package com.devexperts.timetest;
+package com.devexperts.timetest.transformer;
 
 /*
  * #%L